move troop utils to the correct folders
