+++ src/features/Risk/types/troops.ts
@@ -1,0 +1,7 @@
+import { IconType } from 'react-icons';
+
+export interface RiskTroop {
+  Icon: IconType;
+  name: string;
+  value: number;
+}
